% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_stripe_endpoints}
\alias{get_stripe_endpoints}
\title{Get Stripe API Endpoints}
\usage{
get_stripe_endpoints()
}
\value{
A character vector containing all Stripe API endpoints
}
\description{
Returns a character vector of all available Stripe API endpoints. Each
endpoint is prefixed with the HTTP method and includes the full path.
}
