% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{in_ci}
\alias{in_ci}
\title{Check for presence in a CI as an environmental variable}
\usage{
in_ci()
}
\value{
\code{boolean}
}
\description{
Looks in the environmental variable CI.  
Function adapted from private function in \code{package::devtools}.
}

