% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_charges.R
\name{list_charges}
\alias{list_charges}
\title{Fetch Charges from Stripe API}
\usage{
list_charges(client, limit = 10L)
}
\arguments{
\item{client}{A \code{Striper} object created by \code{\link[=rstripe]{rstripe()}}.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing charges data.
}
\description{
Retrieves charges data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe charges see:
\url{https://docs.stripe.com/api/charges/}
}

\examples{
\dontrun{
client <- rstripe("test")
test_charges <- list_charges(client)
live_charges <- list_charges(rstripe("live"))
}

}
