% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_customers.R
\name{list_customers}
\alias{list_customers}
\title{Fetch Customers from Stripe API}
\usage{
list_customers(client, limit = 10L)
}
\arguments{
\item{client}{A \code{Rstripe} object created by \code{\link[=rstripe]{rstripe()}}.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing customers data.
}
\description{
Retrieves customers data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe customers, see:
\url{https://docs.stripe.com/api/customers/}
}

\examples{
\dontrun{
client <- rstripe("test")
test_customers <- list_customers(client)
live_customers <- list_customers(rstripe("live"))
}

}
