% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_invoiceitems.R
\name{list_invoice_items}
\alias{list_invoice_items}
\title{Fetch Invoices Items from Stripe API}
\usage{
list_invoice_items(client, limit = 10L)
}
\arguments{
\item{client}{A \code{Striper} object created by \code{\link[=rstripe]{rstripe()}}.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing invoices items data.
}
\description{
Retrieves invoices items data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe invoice items,
see: \url{https://docs.stripe.com/api/invoiceitems/}
}

\examples{
\dontrun{
client <- rstripe("test")
test_items <- list_invoice_items(client)
live_items <- list_invoice_items(rstripe("live"))
}

}
