% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_prices.R
\name{list_prices}
\alias{list_prices}
\title{Fetch Prices from Stripe API}
\usage{
list_prices(client, limit = 10L)
}
\arguments{
\item{client}{A \code{Rstripe} object created by \code{\link[=rstripe]{rstripe()}}.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing prices data.
}
\description{
Retrieves prices data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe prices, see:
\url{https://docs.stripe.com/api/prices/}
}

\examples{
\dontrun{
client <- rstripe("test")
test_prices <- list_prices(client)
live_prices <- list_prices(rstripe("live"))
}

}
