% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_stripe_key.R
\name{set_stripe_public_key}
\alias{set_stripe_public_key}
\alias{set_stripe_secret_key}
\title{Set Stripe API Keys}
\usage{
set_stripe_public_key(api_key)

set_stripe_secret_key(api_key)
}
\arguments{
\item{api_key}{The public/secret keys to be set}
}
\description{
Sets system environment variables for
 \code{stripe_public_key} and \code{stripe_secret_key}
}
