% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{stripe_api_key}
\alias{stripe_api_key}
\title{Fetch the Stripe API key}
\usage{
stripe_api_key(quiet = getOption("stripe.quiet", FALSE))
}
\arguments{
\item{quiet}{Defaults to FALSE.  Can be set using the R option \code{stripe.quiet}}
}
\value{
\code{character} The API key
}
\description{
Looks in the environmental variable STRIPE_API_KEY.  
Function adapted from \code{github_pat()} in \code{package::devtools}.
}
\examples{
stripe_api_key()
}

