% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_balance_history}
\alias{stripe_balance_history}
\title{Retrieve a Stripe Account Balance History.}
\usage{
stripe_balance_history(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{optional list which can include
\href{https://stripe.com/docs/api#balance_history}{everything here}}
}
\value{
A data frame with balance history information if succeeded.
}
\description{
Retrieve a Stripe Account Balance History.
}

