% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{stripe_cancel_transfer}
\alias{stripe_cancel_transfer}
\title{Cancel a transfer on Stripe.}
\usage{
stripe_cancel_transfer(transfer_id, api_key = NULL)
}
\arguments{
\item{transfer_id}{The id for the tranfer you want to cancel}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with nothing if successful.
}
\description{
This will cancel a transfer that is scheduled.
}
