% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_app_fee_refund}
\alias{stripe_create_app_fee_refund}
\title{Create a refund for an Application Fee.}
\usage{
stripe_create_app_fee_refund(api_key, app_fee_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{app_fee_id}{The id for the Application Fee
you want to refund}

\item{args}{An optional list which can contain an amount to refund
\emph{default is entire fee} and metadata}
}
\value{
A data frame with refund information
}
\description{
Create a refund for an Application Fee.
}
\examples{
stripe_create_app_fee_refund("sk_test_BQokikJOvBiI2HlWgH4olfQ2", "fee_4cbC9iLv8PdUnk")
}

