% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_card_token}
\alias{stripe_create_card_token}
\title{Create a Card Token}
\usage{
stripe_create_card_token(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain card information including
\strong{number}, \strong{exp_month}, and \strong{exp_year}
It can also contain \strong{customer}.}
}
\value{
A data frame with the new card token info if successful.
}
\description{
Create a Card Token
}
\examples{
stripe_create_card_token("sk_test_BQokikJOvBiI2HlWgH4olfQ2",
list(card=list(number="4242424242424242", exp_month="03", exp_year="2018")))
}

