% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_charge}
\alias{stripe_create_charge}
\title{Create a charge.}
\usage{
stripe_create_charge(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain \strong{amount} and
\strong{currency}. It also either needs \strong{customer} or \strong{card}
It can also contain \strong{description}, \strong{metadata},
\strong{capture}
\strong{statement_description}, \strong{receipt_email},
\strong{application_fee}}
}
\value{
A data frame with the new charge info if successful.
}
\description{
Create a new charge for a customer or standalone on card.
}
\examples{
stripe_create_charge("sk_test_BQokikJOvBiI2HlWgH4olfQ2",
list(card="tok_14SsjM2eZvKYlo2CCtdx2AvO", amount=3000, currency="usd",
metadata=list(order_id="12545")))
}

