% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_customer}
\alias{stripe_create_customer}
\title{Create a Customer.}
\usage{
stripe_create_customer(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list which can contain
\strong{card}, \strong{account_balance}, \strong{coupon}, \strong{description},
\strong{metadata}, \strong{email}, \strong{plan}, \strong{quantity},
\strong{trial_end}}
}
\value{
A data frame with the new Customer info if successful.
}
\description{
Create a new customer and add information to them.
}

