% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_invoice}
\alias{stripe_create_invoice}
\title{Create an Invoice.}
\usage{
stripe_create_invoice(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain \strong{customer}
It can also contain \strong{description}, \strong{metadata},
\strong{application_fee}, \strong{subscription},
and \strong{statement_description}}
}
\value{
A data frame with the new invoice if successful.
}
\description{
Create a new invoice for a customer.
}

