% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_invoice_item}
\alias{stripe_create_invoice_item}
\title{Create an Invoice Item.}
\usage{
stripe_create_invoice_item(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain \strong{amount} and
\strong{currency}.
It can also contain \strong{description}, \strong{metadata},
\strong{invoice}, \strong{subscription}.  If you leave invoice blank
it will be added to the next upcoming invoice}
}
\value{
A data frame with the new invoice item if successful.
}
\description{
Create a new invoice item with amount and attach it to a subscription or invoice.
}

