% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_plan}
\alias{stripe_create_plan}
\title{Create a Plan}
\usage{
stripe_create_plan(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain \strong{id}, \strong{amount},
\strong{currency}, \strong{interval}, and \strong{name}.
It can also contain \strong{description}, \strong{metadata},
\strong{interval_count}, \strong{trial_period_days},
and \strong{statement_description}}
}
\value{
A data frame with the new plan info if successful.
}
\description{
Create a new plan on your stripe account.
}
\examples{
stripe_create_plan("sk_test_BQokikJOvBiI2HlWgH4olfQ2",
list(id="example-gold", amount=3000, currency="usd", interval="month"))
}

