% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_recipient}
\alias{stripe_create_recipient}
\title{Add a recipient to Stripe.}
\usage{
stripe_create_recipient(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A required list that must contain \strong{name}
and \strong{type}. It can also contain \strong{tax_id},
\strong{bank_account}, \strong{card}, \strong{email},
\strong{description}, and/or \strong{metadata}.}
}
\value{
A data frame with new recipient information
}
\description{
Add a recipient to Stripe.
}

