% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_refund}
\alias{stripe_create_refund}
\title{Create a Refund}
\usage{
stripe_create_refund(api_key, charge_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The id of the charge to refund.}

\item{args}{A list can contain \strong{amount},
\strong{refund_application_fee}, and \strong{metadata}}
}
\value{
A data frame with the new refund info if successful.
}
\description{
Create a Refund
}

