% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_subscription}
\alias{stripe_create_subscription}
\title{Create a Subscription}
\usage{
stripe_create_subscription(api_key, customer_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The customer id you want to add a subscription to.}

\item{args}{A list which must contain \strong{plan}.
It can also contain \strong{trial_end}, \strong{metadata},
\strong{coupon}, \strong{card}, \strong{application_fee_percent},
and \strong{quantity}}
}
\value{
A data frame with the new subscription info if successful.
}
\description{
Create a new subscription for a customer and add it to them.
}

