% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_transfer}
\alias{stripe_create_transfer}
\title{Create a transfer to Stripe.}
\usage{
stripe_create_transfer(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain \strong{amount},
\strong{currency}, and \strong{recipient}.
It can also contain \strong{description}, \strong{metadata},
\strong{bank_account}, \strong{card},
and \strong{statement_description}}
}
\value{
A data frame with transfer information
}
\description{
Create a transfer to Stripe.
}

