% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{stripe_delete_card}
\alias{stripe_delete_card}
\title{Delete a credit card on Stripe.}
\usage{
stripe_delete_card(customer_id, card_id, api_key = NULL)
}
\arguments{
\item{customer_id}{The id for the customer where the card is.}

\item{card_id}{The id for the card you want to delete}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with card information
}
\description{
Delete a credit card off a customer's account.
}
