% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupons.R
\name{stripe_delete_coupon}
\alias{stripe_delete_coupon}
\title{Delete a coupon.}
\usage{
stripe_delete_coupon(coupon_id, api_key = NULL)
}
\arguments{
\item{coupon_id}{The coupon you want to delete}

\item{api_key}{Your Stripe API Key}
}
\value{
A null dataframe if succeeded.
}
\description{
Delete a specific coupon from your account.
}
