% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{stripe_delete_customer}
\alias{stripe_delete_customer}
\title{Delete a Customer.}
\usage{
stripe_delete_customer(customer_id, api_key = NULL)
}
\arguments{
\item{customer_id}{The id of the customer you want to delete}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame empty if successful.
}
\description{
Remove a customer from your stripe account.
}
