% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounts.R
\name{stripe_delete_customer_discount}
\alias{stripe_delete_customer_discount}
\title{Delete a customer discount.}
\usage{
stripe_delete_customer_discount(customer_id, api_key = NULL)
}
\arguments{
\item{customer_id}{The identifier of the customer who
has the discount to be deleted.}

\item{api_key}{Your Stripe API Key}
}
\value{
Returns a data frame deleted true if if succeeded.
}
\description{
Delete a discout from a customer.
}
