% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice_items.R
\name{stripe_delete_invoice_item}
\alias{stripe_delete_invoice_item}
\title{Delete an Invoice Item.}
\usage{
stripe_delete_invoice_item(item_id, api_key = NULL)
}
\arguments{
\item{item_id}{The invoice item id you want to delete}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame without errors if successful.
}
\description{
Remove an invoice item from an invoice.
}
