% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{stripe_delete_recipient}
\alias{stripe_delete_recipient}
\title{Delete a recipient on Stripe.}
\usage{
stripe_delete_recipient(recipient_id, api_key = NULL)
}
\arguments{
\item{recipient_id}{The id for the recipient you want to delete}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with recipient information
}
\description{
Remove a recipient from your stripe account.
}
