% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripe_key.R
\name{stripe_public_key}
\alias{stripe_public_key}
\alias{stripe_secret_key}
\alias{stripe_public_key_set}
\alias{stripe_secret_key_set}
\title{Stripe API Key Extractors}
\usage{
stripe_public_key()

stripe_secret_key()

stripe_public_key_set()

stripe_secret_key_set()
}
\value{
Character if the key exists, \code{NULL} otherwise.
}
\description{
Grabs system environment variables (e.g. from ~/.Renviron) if
they are available.  For \code{stripe_public_key}, the environment variable must
be \code{stripe_public_key} and \code{stripe_secret_key} it must be
\code{stripe_secret_key}.  \code{stripe_secret/public_key_set} are simple
tests to see if the keys are set
}
