% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_application_fee_refunds}
\alias{stripe_list_application_fee_refunds}
\title{List all refunds for an Application Fee.}
\usage{
stripe_list_application_fee_refunds(api_key, app_fee_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{app_fee_id}{The id for the Application Fee
you want to refund}

\item{args}{An optional list which can contain
\strong{ending_before}, \strong{limit}, or \strong{starting_after}}
}
\value{
A data frame with application fee refund object if succeeded
}
\description{
List all refunds for an Application Fee.
}
\examples{
stripe_list_application_fee_refunds("sk_test_BQokikJOvBiI2HlWgH4olfQ2", "fee_4cbC9iLv8PdUnk",
list(limit=3)) #returns 3 refunds
}

