% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_application_fees}
\alias{stripe_list_application_fees}
\title{List all Application Fees.}
\usage{
stripe_list_application_fees(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{optional list that can include
\strong{charge}, \strong{created}, \strong{ending_before},
\strong{limit}, \strong{starting_after}}
}
\value{
A data frame with the application fees you
have created that allow you to refund them.
}
\description{
List all the application fees.  Can filter by charge, date, etc.
}

