% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_cards}
\alias{stripe_list_cards}
\title{List credit card on customers.}
\usage{
stripe_list_cards(api_key, customer_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The id for the customer you want to list the cards}

\item{args}{an optional list that can contain
/strong{ending_before}, \strong{limit}, \strong{starting_after}}
}
\value{
A data frame with card information
}
\description{
List all the credit cards linked to a customer.
}

