% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_charges}
\alias{stripe_list_charges}
\title{List all charges.}
\usage{
stripe_list_charges(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list that can include \strong{customer},
\strong{created}, \strong{optional}, \strong{ending_before},
\strong{limit}, \strong{starting_after}}
}
\value{
Returns a data frame with all customer charges.  10 if no limit specified.
}
\description{
List all charges.
}

