% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupons.R
\name{stripe_list_coupons}
\alias{stripe_list_coupons}
\title{List all coupons.}
\usage{
stripe_list_coupons(args = NULL, api_key = NULL)
}
\arguments{
\item{args}{An optional list which can contain
\describe{
   \item{ending_before}{\emph{optional:} An object id which will show
   objects before}
   \item{limit}{\emph{optional:} A number 1 to 100 to limit the items.
   Default is 10}
   \item{starting_after}{\emph{optional:} An object id which will show
   objects starting here}
}}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with all the coupons.
}
\description{
List all the coupons currently in your account.
}
