% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_customers}
\alias{stripe_list_customers}
\title{Retrieve a list of Customer.}
\usage{
stripe_list_customers(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list of arguments. Can include
\strong{created}, \strong{ending_before}, \strong{limit}
\strong{starting_after}. Limit defaults at 10.}
}
\value{
A data frame with the Customer list if successful.
}
\description{
Retrieve a list of Customer.
}

