% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_events}
\alias{stripe_list_events}
\title{List all events.}
\usage{
stripe_list_events(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list that can include \strong{type},
\strong{created}, \strong{ending_before},
\strong{limit}, \strong{starting_after}}
}
\value{
Returns an event data frame containing
all the eventsif valid.
}
\description{
list all the events that have happened.
}

