% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_invoice_items}
\alias{stripe_list_invoice_items}
\title{List all Invoice Item.}
\usage{
stripe_list_invoice_items(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain \strong{created},
\strong{customer}, \strong{ending_before}, \strong{limit}
and \strong{starting_after}}
}
\value{
A data frame with the the invoice items if successful.
}
\description{
List all the invoice items in your account.  This can be limited by customer, date, etc.
}

