% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_plans}
\alias{stripe_list_plans}
\title{List Plans.}
\usage{
stripe_list_plans(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list that can contain
/strong{ending_before}, \strong{limit}, \strong{starting_after}}
}
\value{
A data frame with plan information
}
\description{
List Plans.
}

