% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_refunds}
\alias{stripe_list_refunds}
\title{List Refunds}
\usage{
stripe_list_refunds(api_key, charge_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The id of the charge with the refund.}

\item{args}{an optional list that can contain
/strong{ending_before}, \strong{limit}, \strong{starting_after}}
}
\value{
A data frame with refunds information
}
\description{
List Refunds
}

