% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_subscriptions}
\alias{stripe_list_subscriptions}
\title{List Subscriptions}
\usage{
stripe_list_subscriptions(api_key, customer_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The id of the customer with the subscriptions.}

\item{args}{an optional list that can contain
/strong{ending_before}, \strong{limit}, \strong{starting_after}}
}
\value{
A data frame with subscriptions information
}
\description{
List Subscriptions
}

