% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_transfers}
\alias{stripe_list_transfers}
\title{List Transfers.}
\usage{
stripe_list_transfers(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list that can contain \strong{date}
\strong{ending_before}, \strong{limit}, \strong{starting_after},
\strong{recipient}, and \strong{status}}
}
\value{
A data frame with transfers information
}
\description{
List all the Bank Accoun transfers for the account.
}

