% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_pay_invoice}
\alias{stripe_pay_invoice}
\title{Pay an Invoice.}
\usage{
stripe_pay_invoice(invoice_id, api_key = NULL)
}
\arguments{
\item{invoice_id}{The invoice id you want to pay}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with the invoice if successful.
}
\description{
Pay the invoice right now with customer information.
}
