% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_account}
\alias{stripe_retrieve_account}
\title{Your Stripe Account.}
\usage{
stripe_retrieve_account(api_key)
}
\arguments{
\item{api_key}{Your Stripe API key}
}
\value{
Your account information in a data frame.
}
\description{
Retrieve your account.
}
\examples{
stripe_retrieve_account("sk_test_2337VDxzNLS332zsso")
}

