% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_app_fee}
\alias{stripe_retrieve_app_fee}
\title{Retrieve an App Fee.}
\usage{
stripe_retrieve_app_fee(api_key, app_fee_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{app_fee_id}{The id for the Application Fee
you want to retrieve}
}
\value{
A data frame with refund app fee information
}
\description{
Retrieve an App Fee.
}

