% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_fees.R
\name{stripe_retrieve_application_fee}
\alias{stripe_retrieve_application_fee}
\title{Retrieve an Application Fee.}
\usage{
stripe_retrieve_application_fee(application_fee_id, api_key = NULL)
}
\arguments{
\item{application_fee_id}{The id for the Application Fee
you want to retrieve}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with refund app fee information
}
\description{
Retrieve the data behind a specific applicaiton fee.
}
