% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_fee_refunds.R
\name{stripe_retrieve_application_fee_refund}
\alias{stripe_retrieve_application_fee_refund}
\title{Retrieve an existing refund for an Application Fee.}
\usage{
stripe_retrieve_application_fee_refund(application_fee_id, refund_id,
  api_key = NULL)
}
\arguments{
\item{application_fee_id}{The id for the Application Fee
you want to refund}

\item{refund_id}{The id for the Refund
you want to retrieve}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with refund information
}
\description{
Retrieve the information of a refund.
}
