% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_balance}
\alias{stripe_retrieve_balance}
\title{Retrieve Stripe Account Balance.}
\usage{
stripe_retrieve_balance(api_key)
}
\arguments{
\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with your account balance
}
\description{
Retrieve Stripe Account Balance.
}

