% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{stripe_retrieve_card}
\alias{stripe_retrieve_card}
\title{Retrieve a card from a customer}
\usage{
stripe_retrieve_card(card_id, customer_id, api_key = NULL)
}
\arguments{
\item{card_id}{The id for the card you want to retrieve}

\item{customer_id}{The id for the customer where the card is
added to.}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with card information
}
\description{
Retrieve the information about a credit card.  Can retrieve last 4 digits.
}
