% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_coupon}
\alias{stripe_retrieve_coupon}
\title{Retrieve a coupon.}
\usage{
stripe_retrieve_coupon(api_key, coupon_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{coupon_id}{The coupon you want to retrieve.}
}
\value{
A data frame with the coupon information
}
\description{
Retrieve a coupon.
}

