% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{stripe_retrieve_customer}
\alias{stripe_retrieve_customer}
\title{Retrieve a Customer.}
\usage{
stripe_retrieve_customer(api_key, customer_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The id of the customer you want to retrieve}
}
\value{
A data frame with the Customer info if successful.
}
\description{
Retrieve the information about a specific customer using their id.
}

