% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{stripe_retrieve_event}
\alias{stripe_retrieve_event}
\title{Retrieve an event.}
\usage{
stripe_retrieve_event(event_id, api_key = NULL)
}
\arguments{
\item{event_id}{The identifier of the event to be retrieved}

\item{api_key}{Your Stripe API Key}
}
\value{
Returns an event data frame if valid.
}
\description{
Retrieve the information about a specific event.
}
