% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_retrieve_invoice}
\alias{stripe_retrieve_invoice}
\title{Retrieve an Invoice.}
\usage{
stripe_retrieve_invoice(invoice_id, api_key = NULL)
}
\arguments{
\item{invoice_id}{The invoice id you want to retrieve}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with the invoice if successful.
}
\description{
Retrieve all the information on a particular invoice.
}
