% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_invoice_line_items}
\alias{stripe_retrieve_invoice_line_items}
\title{Retrieve Invoice Items.}
\usage{
stripe_retrieve_invoice_line_items(api_key, invoice_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{invoice_id}{The invoice id with the items you want to retrieve}

\item{args}{An optional list that can include \strong{limit}, \strong{customer}
\strong{ending_before}, \strong{subscription}, and/or \strong{starting_after}}
}
\value{
A data frame with the invoice items if successful.
}
\description{
Retreive all the items on an invoice.
}

