% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{stripe_retrieve_recipient}
\alias{stripe_retrieve_recipient}
\title{Retrieve information on a recipient.}
\usage{
stripe_retrieve_recipient(recipient_id, api_key = NULL)
}
\arguments{
\item{recipient_id}{The id for the recipient you want to retrieve.}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with recipient information
}
\description{
Retrieve all the information on a recipient.
}
