% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_refund}
\alias{stripe_retrieve_refund}
\title{Retrieve a refund.}
\usage{
stripe_retrieve_refund(api_key, charge_id, refund_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The id of the charge with the refund.}

\item{refund_id}{The id of the refund you want to retrieve}
}
\value{
A data frame with the refund info if successful.
}
\description{
Retrieve a refund.
}

