% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_transfer}
\alias{stripe_retrieve_transfer}
\title{Retrieve a transfer}
\usage{
stripe_retrieve_transfer(api_key, transfer_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{transfer_id}{The id of the transfer you want to retrieve}
}
\value{
A data frame with the transfer info if successful.
}
\description{
Retrieve a transfer
}

