% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_app_fee_refund}
\alias{stripe_update_app_fee_refund}
\title{Update a refund for an Application Fee.}
\usage{
stripe_update_app_fee_refund(api_key, app_fee_id, refund_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{app_fee_id}{The id for the Application Fee
you want to refund}

\item{refund_id}{The id for the Refund
you want to update}

\item{args}{An optional list which can contain an amount to refund
\emph{default is entire fee} and metadata}
}
\value{
A data frame with application fee refund object if succeeded
}
\description{
Update a refund for an Application Fee.
}
\examples{
stripe_create_app_fee_refund("sk_test_BQokikJOvBiI2HlWgH4olfQ2", "fee_4cbC9iLv8PdUnk",
list(metadata=list(reason="Accidental charge")))
}

