% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_fee_refunds.R
\name{stripe_update_application_fee_refund}
\alias{stripe_update_application_fee_refund}
\title{Update a refund for an Application Fee.}
\usage{
stripe_update_application_fee_refund(api_key, application_fee_id, refund_id,
  args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{application_fee_id}{The id for the Application Fee
you want to refund}

\item{refund_id}{The id for the Refund 
you want to update}

\item{args}{An optional list which can contain:
\describe{
   \item{metadata}{\emph{optional} A list which can contain anything to further describe.}
}}
}
\value{
A data frame with application fee refund object if succeeded
}
\description{
Update all the information of a refund of an application fee.
}
\examples{
\dontrun{stripe_create_application_fee_refund("sk_test_BQokikJOvBiI2HlWgH4olfQ2", "fee_4cbC9iLv8PdUnk",
list(metadata=list(reason="Accidental charge")))}

}

