% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_coupon}
\alias{stripe_update_coupon}
\title{Update a coupon.}
\usage{
stripe_update_coupon(api_key, coupon_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{coupon_id}{The coupon you want to update.}

\item{args}{An list which can contain a \emph{list of \strong{metadata}}
and/or a \strong{description}}
}
\value{
A data frame with the new coupon information if succeeded.
}
\description{
Update a coupon.
}

