% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_customer}
\alias{stripe_update_customer}
\title{Update a Customer.}
\usage{
stripe_update_customer(api_key, customer_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The id of the customer you want to update}

\item{args}{an optional list which can contain
\strong{card}, \strong{account_balance}, \strong{coupon}, \strong{description},
\strong{metadata}, \strong{email}, \strong{default_card},}
}
\value{
A data frame with the updated Customer info if successful.
}
\description{
Update a Customer.
}

