% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_update_invoice}
\alias{stripe_update_invoice}
\title{Update an Invoice.}
\usage{
stripe_update_invoice(invoice_id, args, api_key = NULL)
}
\arguments{
\item{invoice_id}{The invoice id you want to update}

\item{args}{Can include:
\describe{
   \item{application_fee}{}
   \item{description}{}
   \item{closed}{}
   \item{forgiven}{}
   \item{statement_description}{}
   \item{metadata}{}
}}

\item{api_key}{Your Stripe API Key}
}
\value{
A data frame with the updated invoice if successful.
}
\description{
Update the information on an invoice, add a fee, close, or forgive it.
}
