% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_invoice_item}
\alias{stripe_update_invoice_item}
\title{Update an Invoice Item.}
\usage{
stripe_update_invoice_item(api_key, item_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{item_id}{The invoice item id you want to update}

\item{args}{Can include \strong{amount}, \strong{description}
and/or \strong{metadata}}
}
\value{
A data frame with the updated invoice item if successful.
}
\description{
Update an Invoice Item.
}

