% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_recipient}
\alias{stripe_update_recipient}
\title{Update a recipient on Stripe.}
\usage{
stripe_update_recipient(api_key, recipient_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{recipient_id}{The id for the recipient you want to update.}

\item{args}{A list that can contain \strong{name}, \strong{tax_id},
\strong{bank_account}, \strong{card}, \strong{default_card}, \strong{email},
\strong{description}, and/or \strong{metadata}.}
}
\value{
A data frame with recipient information
}
\description{
Update a recipient on Stripe.
}

