% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_refund}
\alias{stripe_update_refund}
\title{Update a Refund}
\usage{
stripe_update_refund(api_key, charge_id, refund_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The id of the charge with the refund.}

\item{refund_id}{The id of the refund you want to update}

\item{args}{A list can contain \strong{metadata}.}
}
\value{
A data frame with the updated refund info if successful.
}
\description{
Update a Refund
}

