% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_subscription}
\alias{stripe_update_subscription}
\title{Update a subscription.}
\usage{
stripe_update_subscription(api_key, customer_id, subscription_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The id of the customer with the subscription.}

\item{subscription_id}{The id of the subscription you want to update.}

\item{args}{A list can contain \strong{plan},
\strong{coupon}, \strong{prorate}, \strong{trial_end},
\strong{application_fee_percent}, \strong{quantity},
\strong{metadata}, and \strong{card}}
}
\value{
A data frame with the updated subscription info if successful.
}
\description{
Update a subscription.
}

