% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_transfer}
\alias{stripe_update_transfer}
\title{Update a Transfer}
\usage{
stripe_update_transfer(api_key, transfer_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{transfer_id}{The id of the transfer you want to update}

\item{args}{A list can contain
\strong{metadata}, and \strong{description}}
}
\value{
A data frame with the updated transfer info if successful.
}
\description{
You can update a transfer that is scheduled or has been accepted by changing description.
}

