% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstripe-package.R
\docType{package}
\name{rstripe-package}
\alias{rstripe}
\alias{rstripe-package}
\title{rstripe: R Interface to the Stripe API}
\description{
Provides a R interface to interact with the Stripe API. Includes functions for fetching and processing various Stripe resources such as transactions, customers, products, and subscriptions.
}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea@nutrivetpet.com}

}
\keyword{internal}
