% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_balance_transactions.R
\name{list_balance_transactions}
\alias{list_balance_transactions}
\title{Fetch Balance Transactions from Stripe API}
\usage{
list_balance_transactions(client, limit = 10L)
}
\arguments{
\item{client}{A \code{Rstripe} object created by \code{\link[=rstripe]{rstripe()}}.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing balance transaction
data.
}
\description{
Retrieves balance transaction data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe balance
transactions, see: \url{https://docs.stripe.com/api/balance_transactions/}
}

\examples{
\dontrun{
client <- rstripe("test")
test_transactions <- list_balance_transactions(client)
live_transactions <- list_balance_transactions(rstripe("live"))
}

}
\concept{endpoints}
