% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charges.R
\name{list_charges}
\alias{list_charges}
\title{Fetch Charges from Stripe API}
\usage{
list_charges(mode = c("test", "live"), limit = 10L)
}
\arguments{
\item{mode}{Character string specifying the API mode. Must be either "test"
for test mode or "live" for live mode. The API key will be automatically
retrieved based on this mode.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing charges
data.
}
\description{
Retrieves charges data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe charges see:
\url{https://docs.stripe.com/api/charges/}
}

\examples{
\dontrun{
# Fetch test mode balance transactions
test_charges <- list_charges("test")

# Fetch live mode balance transactions
live_charges <- list_charges("live")
}

}
