% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_invoices.R
\name{list_invoices}
\alias{list_invoices}
\title{Fetch Invoices from Stripe API}
\usage{
list_invoices(client, limit = 10L)
}
\arguments{
\item{client}{A \code{Striper} object created by \code{\link[=rstripe]{rstripe()}}.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing invoices data.
}
\description{
Retrieves invoices data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe invoices, see:
\url{https://docs.stripe.com/api/invoices/}
}

\examples{
\dontrun{
client <- rstripe("test")
test_invoices <- list_invoices(client)
live_invoices <- list_invoices(rstripe("live"))
}

}
