% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_products.R
\name{list_products}
\alias{list_products}
\title{Fetch Products from Stripe API}
\usage{
list_products(client, limit = 10L)
}
\arguments{
\item{client}{A \code{Striper} object created by \code{\link[=rstripe]{rstripe()}}.}

\item{limit}{Either a scalar between 1 and 100 or \code{Inf} to traverse all the
available data. 10 by default.}
}
\value{
A data frame (tibble if available) containing products data.
}
\description{
Retrieves products data from the Stripe API.
}
\section{API Documentation}{
 For more information about Stripe products, see:
\url{https://docs.stripe.com/api/products/}
}

\examples{
\dontrun{
client <- rstripe("test")
test_products <- list_products(client)
live_products <- list_products(rstripe("live"))
}

}
