% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_capture_charge}
\alias{stripe_capture_charge}
\title{Capture a charge.}
\usage{
stripe_capture_charge(api_key, charge_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The identifier of the charge to be captured}

\item{args}{an optional list that can include \strong{amount},
\strong{application_fee}, \strong{receipt_email}}
}
\value{
Returns a charge data frame if the capture succeeded.
}
\description{
As part of the two step charge workflow, capture a charge.  Read more about the workflow
on the stripe website.
}

