% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_account_token}
\alias{stripe_create_account_token}
\title{Create a Bank Account Token}
\usage{
stripe_create_account_token(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain bank account information including
\strong{country}, \strong{routing_number}, and \strong{account_number}}
}
\value{
A data frame with the new bank account token info if successful.
}
\description{
Create a Bank Account Token
}
\examples{
stripe_create_account_token("sk_test_BQokikJOvBiI2HlWgH4olfQ2",
list(card=list(country="US", routing_number="110000000", account_number="0000000014343")))
}

