% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_fee_refunds.R
\name{stripe_create_application_fee_refund}
\alias{stripe_create_application_fee_refund}
\title{Create a refund for an Application Fee.}
\usage{
stripe_create_application_fee_refund(api_key, application_fee_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{application_fee_id}{The id for the Application Fee
you want to refund}

\item{args}{An optional list which can contain an amount to refund
\describe{
   \item{amount}{\emph{optional} default is entire fee} 
   \item{metadata}{\emph{optional} A list which can contain anything to further describe.}
}}
}
\value{
A data frame with refund information
}
\description{
Create a refund for an application fee.
}
\examples{
\dontrun{stripe_create_application_fee_refund("sk_test_BQokikJOvBiI2HlWgH4olfQ2", "fee_4cbC9iLv8PdUnk")}

}

