% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_create_coupon}
\alias{stripe_create_coupon}
\title{Create a coupon.}
\usage{
stripe_create_coupon(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{An list which must contain \strong{duration}.
It can also contain \strong{id}, \strong{amount_off},
\strong{currency} and more.  View \href{htts://stripe.com/docs/api#create_coupon}{here}.}
}
\value{
A data frame with the coupon information
}
\description{
Create a coupon.
}
\examples{
stripe_create_coupon("sk_test_BQokikJOvBiI2HlWgH4olfQ2",
list(amount_off=200, duration='forever'))
}

