% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{stripe_create_customer}
\alias{stripe_create_customer}
\title{Create a Customer.}
\usage{
stripe_create_customer(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list which can contain
\describe{
   \item{account_balance}{\emph{optional} In cents starting account balance. Positive means
                          amount to be paid off. Negative is credit used towards charge.}
   \item{card}{\emph{optional} Either token from Stripe.js or card details.}
   \item{description}{\emph{optional} The description of the transfer.}
   \item{metadata}{\emph{optional} A list which can contain anything to further describe.}
   \item{coupon}{\emph{optional} A coupon code for a discount applied on all charges.}
   \item{email}{\emph{optional} Customer's email address.}
   \item{plan}{\emph{optional} A plan id to subscribe the customer to.}
   \item{quantity}{\emph{optional} The amount of plans to subscribe them to.} 
   \item{trial_end}{\emph{optional} Timestamp for end of trial period. Overrides plan trial.}
}}
}
\value{
A data frame with the new Customer info if successful.
}
\description{
Create a new customer and add information to them.
}

