% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_create_invoice}
\alias{stripe_create_invoice}
\title{Create an Invoice.}
\usage{
stripe_create_invoice(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which must contain 
\describe{
   \item{customer}{\strong{required} customer id for the invoice.}
   \item{description}{\emph{optional} description of invoice.}
   \item{application_fee}{\emph{optional} fee in cents to apply to invoice.}
   \item{metadata}{\emph{optional} A list which can contain anything to further describe.}
   \item{subscription}{\emph{optional} id of the subscription to invoice.}
   \item{statement_description}{\emph{optional} Goes on credit card statement of customer.}
}}
}
\value{
A data frame with the new invoice if successful.
}
\description{
Create a new invoice for a customer.
}

