% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{stripe_create_recipient}
\alias{stripe_create_recipient}
\title{Add a recipient to Stripe.}
\usage{
stripe_create_recipient(api_key, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A required list that must contain 
\describe{ 
   \item{name}{\strong{required} The legal name of recipient.}
   \item{type}{\strong{required} put 'individual' or 'corporation'.}
   \item{tax_id}{\emph{optional} The tax ID either 'SSN' for individual or 'EIN' for corporation}
   \item{bank_account}{\emph{optional} token or dictionary of details.}
   \item{card}{\emph{optional} A US Visa/Mastercard debit card to put transfer on.}
   \item{email}{\emph{optional} Recipients email address.}
   \item{description}{\emph{optional} Arbitrary string to describe.}
   \item{metadata}{\emph{optional} key/value for storing information.}
}}
}
\value{
A data frame with new recipient information
}
\description{
Add a new recipient to your account to schedule transfers and more.
}

