% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupons.R
\name{stripe_delete_coupon}
\alias{stripe_delete_coupon}
\title{Delete a coupon.}
\usage{
stripe_delete_coupon(api_key, coupon_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{coupon_id}{The coupon you want to delete}
}
\value{
A null dataframe if succeeded.
}
\description{
Delete a specific coupon from your account.
}

