% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_delete_customer}
\alias{stripe_delete_customer}
\title{Delete a Customer.}
\usage{
stripe_delete_customer(api_key, customer_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The id of the customer you want to delete}
}
\value{
A data frame empty if successful.
}
\description{
Delete a Customer.
}

