% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounts.R
\name{stripe_delete_customer_discount}
\alias{stripe_delete_customer_discount}
\title{Delete a customer discount.}
\usage{
stripe_delete_customer_discount(api_key, customer_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The identifier of the customer who
has the discount to be deleted.}
}
\value{
Returns a data frame deleted true if if succeeded.
}
\description{
Delete a discout from a customer.
}

