% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_delete_invoice_item}
\alias{stripe_delete_invoice_item}
\title{Delete an Invoice Item.}
\usage{
stripe_delete_invoice_item(api_key, item_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{item_id}{The invoice item id you want to delete}
}
\value{
A data frame without errors if successful.
}
\description{
Remove an invoice item from an invoice.
}

