% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_delete_recipient}
\alias{stripe_delete_recipient}
\title{Delete a recipient on Stripe.}
\usage{
stripe_delete_recipient(api_key, recipient_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{recipient}{The id for the recipient you want to delete}
}
\value{
A data frame with recipient information
}
\description{
Delete a recipient on Stripe.
}

