% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounts.R
\name{stripe_delete_subscription_discount}
\alias{stripe_delete_subscription_discount}
\title{Delete a subscription discount.}
\usage{
stripe_delete_subscription_discount(api_key, customer_id, subscription_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The identifier of the customer who
has the discount to be deleted.}

\item{subscription_id}{The identifier of the subscription.}
}
\value{
Returns a data frame deleted true if if succeeded.
}
\description{
Delete a discount from a customer's subscription.
}

