% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_fees.R
\name{stripe_list_application_fees}
\alias{stripe_list_application_fees}
\title{List all Application Fees.}
\usage{
stripe_list_application_fees(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{optional list that can include:
\describe{
   \item{charge}{\emph{optional} Only return feew associated with this charge.}
   \item{ending_before}{\emph{optional:} An object id which will show objects before}
   \item{limit}{\emph{optional:} A number 1 to 100 to limit the items.  Default is 10}
   \item{starting_after}{\emph{optional:} An object id which will show objects starting here}
}}
}
\value{
A data frame with the application fees you
have created that allow you to refund them.
}
\description{
List all the application fees.  Can filter by charge, date, etc.
}

