% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_coupons}
\alias{stripe_list_coupons}
\title{List all coupons.}
\usage{
stripe_list_coupons(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{An optional list which can contain
\strong{ending_before}, \strong{limit}, or \strong{starting_after}}
}
\value{
A data frame with all the coupons.
}
\description{
List all coupons.
}
\examples{
stripe_list_application_fee_refunds("sk_test_BQokikJOvBiI2HlWgH4olfQ2", "fee_4cbC9iLv8PdUnk",
list(limit=3)) #returns 3 refunds
}

