% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{stripe_list_customers}
\alias{stripe_list_customers}
\title{Retrieve a list of Customer.}
\usage{
stripe_list_customers(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list of arguments. Can include
\describe{
   \item{ending_before}{\emph{optional:} An object id which will show objects before}
   \item{limit}{\emph{optional:} A number 1 to 100 to limit the items.  Default is 10}
   \item{starting_after}{\emph{optional:} An object id which will show objects starting here}
}}
}
\value{
A data frame with the Customer list if successful.
}
\description{
Retrieve a list of all customers.  Defaults up to 10.
}

