% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoice_items.R
\name{stripe_list_invoice_items}
\alias{stripe_list_invoice_items}
\title{List all Invoice Item.}
\usage{
stripe_list_invoice_items(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{A list which can contain:
\describe{
   \item{ending_before}{\emph{optional:} An object id which will show objects before}
   \item{customer}{\emph{optional:} The customer ID whose invoice items to return}
   \item{limit}{\emph{optional:} A number 1 to 100 to limit the items.  Default is 10}
   \item{starting_after}{\emph{optional:} An object id which will show objects starting here}
}}
}
\value{
A data frame with the the invoice items if successful.
}
\description{
List all the invoice items in your account.  This can be limited by customer, date, etc.
}

