% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_invoices}
\alias{stripe_list_invoices}
\title{List Invoices.}
\usage{
stripe_list_invoices(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{An optional list that can include \strong{customer},
\strong{date}, \strong{ending_before}, \strong{limit}, \strong{starting_after}}
}
\value{
A data frame with the invoices if successful.
}
\description{
List Invoices.
}

