% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_list_invoices}
\alias{stripe_list_invoices}
\title{List Invoices.}
\usage{
stripe_list_invoices(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{An optional list that can include:
\describe{
   \item{ending_before}{\emph{optional:} An object id which will show objects before}
   \item{limit}{\emph{optional:} A number 1 to 100 to limit the items.  Default is 10}
   \item{starting_after}{\emph{optional:} An object id which will show objects starting here}
}}
}
\value{
A data frame with the invoices if successful.
}
\description{
List invoices for a particular customer or all customers within a date.
}

