% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_list_recipients}
\alias{stripe_list_recipients}
\title{List all current recipients.}
\usage{
stripe_list_recipients(api_key, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{args}{an optional list that can contain \strong{verified}
/strong{ending_before}, \strong{limit}, \strong{starting_after}}
}
\value{
A data frame with recipient's information
}
\description{
List all the current recipients in your account.
}

