% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refunds.R
\name{stripe_list_refunds}
\alias{stripe_list_refunds}
\title{List Refunds}
\usage{
stripe_list_refunds(api_key, charge_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The id of the charge with the refund.}

\item{args}{an optional list that can contain:
\describe{
   \item{ending_before}{\emph{optional:} An object id which will show objects before}
   \item{limit}{\emph{optional:} A number 1 to 100 to limit the items.  Default is 10}
   \item{starting_after}{\emph{optional:} An object id which will show objects starting here}
}}
}
\value{
A data frame with refunds information
}
\description{
List all the refunds that a charge has.  You can have multiple refunds per charge until the 
full amount is reached.
}

