% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charges.R
\name{stripe_retrieve_charge}
\alias{stripe_retrieve_charge}
\title{Retrieve a charge.}
\usage{
stripe_retrieve_charge(api_key, charge_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The identifier of the charge to be retrieved}
}
\value{
Returns a charge data frame if valid.
}
\description{
Retrieve all the information about a specific charge.
}
\examples{
\dontrun{stripe_retrieve_charge("sk_test_BQokikJOvBiI2HlWgH4olfQ2", 
"ch_14TLyd2eZvKYlo2CdryxwrF8")}

}

