% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{stripe_retrieve_event}
\alias{stripe_retrieve_event}
\title{Retrieve an event.}
\usage{
stripe_retrieve_event(api_key, event_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{event_id}{The identifier of the event to be retrieved}
}
\value{
Returns an event data frame if valid.
}
\description{
Retrieve the information about a specific event.
}

