% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_invoice}
\alias{stripe_retrieve_invoice}
\title{Retrieve an Invoice.}
\usage{
stripe_retrieve_invoice(api_key, invoice_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{invoice_id}{The invoice id you want to retrieve}
}
\value{
A data frame with the invoice if successful.
}
\description{
Retrieve an Invoice.
}

