% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_retrieve_invoice}
\alias{stripe_retrieve_invoice}
\title{Retrieve an Invoice.}
\usage{
stripe_retrieve_invoice(api_key, invoice_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{invoice_id}{The invoice id you want to retrieve}
}
\value{
A data frame with the invoice if successful.
}
\description{
Retrieve all the information on a particular invoice.
}

