% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_invoice_item}
\alias{stripe_retrieve_invoice_item}
\title{Retrieve an Invoice Item.}
\usage{
stripe_retrieve_invoice_item(api_key, item_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{item_id}{The invoice item id you want to retrieve}
}
\value{
A data frame with the invoice item if successful.
}
\description{
Retrieve a specific invoice item's info using it's id.
}

