% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_retrieve_invoice_line_items}
\alias{stripe_retrieve_invoice_line_items}
\title{Retrieve Invoice Items.}
\usage{
stripe_retrieve_invoice_line_items(api_key, invoice_id, args = NULL)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{invoice_id}{The invoice id with the items you want to retrieve}

\item{args}{An optional list that can include 
\describe{
   \item{ending_before}{\emph{optional:} An object id which will show objects before}
   \item{limit}{\emph{optional:} A number 1 to 100 to limit the items.  Default is 10}
   \item{starting_after}{\emph{optional:} An object id which will show objects starting here}
}}
}
\value{
A data frame with the invoice items if successful.
}
\description{
Retreive all the items on an invoice.
}

