% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{stripe_retrieve_plan}
\alias{stripe_retrieve_plan}
\title{Retrieve a Plan}
\usage{
stripe_retrieve_plan(api_key, plan_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{plan_id}{The id of the plan you want to retrieve}
}
\value{
A data frame with the plan info if successful.
}
\description{
Retrieve the information on a particular plan on your account.
}

