% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refunds.R
\name{stripe_retrieve_refund}
\alias{stripe_retrieve_refund}
\title{Retrieve a refund.}
\usage{
stripe_retrieve_refund(api_key, charge_id, refund_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The id of the charge with the refund.}

\item{refund_id}{The id of the refund you want to retrieve}
}
\value{
A data frame with the refund info if successful.
}
\description{
Retrieve all the information about a particular refund on a charge.
}

