% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_retrieve_subscription}
\alias{stripe_retrieve_subscription}
\title{Retrieve a Subscription}
\usage{
stripe_retrieve_subscription(api_key, customer_id, subscription_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The id of the customer with the subscription.}

\item{subscription_id}{The id of the subscription you want to retrieve}
}
\value{
A data frame with the subscription info if successful.
}
\description{
Retrieve a Subscription
}

