% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\name{stripe_retrieve_transfer}
\alias{stripe_retrieve_transfer}
\title{Retrieve a transfer}
\usage{
stripe_retrieve_transfer(api_key, transfer_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{transfer_id}{The id of the transfer you want to retrieve}
}
\value{
A data frame with the transfer info if successful.
}
\description{
You can get all the information about a pending transfer or an accepted transfer.
}

