% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{stripe_upcoming_customer_invoice}
\alias{stripe_upcoming_customer_invoice}
\title{Retrieve upcoming customer's invoice.}
\usage{
stripe_upcoming_customer_invoice(api_key, customer_id)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{customer_id}{The customer id that you want the upcoming invoice}
}
\value{
A data frame with the upcoming invoice if successful.
}
\description{
Retrieve the upcoming customer invoice to send out.
}

