% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupons.R
\name{stripe_update_coupon}
\alias{stripe_update_coupon}
\title{Update a coupon.}
\usage{
stripe_update_coupon(api_key, coupon_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{coupon_id}{The coupon you want to update.}

\item{args}{An list which can contain:
\describe{
   \item{description}{\emph{optional} The description of the transfer.}
   \item{metadata}{\emph{optional} A list which can contain anything to further describe.}
}}
}
\value{
A data frame with the new coupon information if succeeded.
}
\description{
Update the information about a coupon.
}

