% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_invoice}
\alias{stripe_update_invoice}
\title{Update an Invoice.}
\usage{
stripe_update_invoice(api_key, invoice_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{invoice_id}{The invoice id you want to update}

\item{args}{Can include \strong{application_fee}, \strong{description}
\strong{closed}, \strong{forgiven}, \strong{statement_description},
and/or \strong{metadata}}
}
\value{
A data frame with the updated invoice if successful.
}
\description{
Update an Invoice.
}

