% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stripe_update_plan}
\alias{stripe_update_plan}
\title{Update a Plan}
\usage{
stripe_update_plan(api_key, plan_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{plan_id}{The id of the plan you want to update}

\item{args}{A list can contain \strong{statement_description},
\strong{metadata}, and \strong{name}}
}
\value{
A data frame with the updated plan info if successful.
}
\description{
Update the information about a particular plan on your stripe account.
}

