% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipients.R
\name{stripe_update_recipient}
\alias{stripe_update_recipient}
\title{Update a recipient on Stripe.}
\usage{
stripe_update_recipient(api_key, recipient_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{recipient_id}{The id for the recipient you want to update.}

\item{args}{A list that can contain 
\describe{
   \item{name}{\emph{optional} The legal name of recipient.}
   \item{tax_id}{\emph{optional} The tax ID either 'SSN' for individual or 'EIN' for corporation}
   \item{bank_account}{\emph{optional} token or dictionary of details.}
   \item{card}{\emph{optional} A US Visa/Mastercard debit card to put transfer on.}
   \item{email}{\emph{optional} Recipients email address.}
   \item{description}{\emph{optional} Arbitrary string to describe.}
   \item{metadata}{\emph{optional} key/value for storing information.}
   \item{default_card}{\emph{optional} The default card to sue.} 
}}
}
\value{
A data frame with recipient information
}
\description{
Update or add information about a recipient on stripe.
}

