% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refunds.R
\name{stripe_update_refund}
\alias{stripe_update_refund}
\title{Update a Refund}
\usage{
stripe_update_refund(api_key, charge_id, refund_id, args)
}
\arguments{
\item{api_key}{Your Stripe API Key}

\item{charge_id}{The id of the charge with the refund.}

\item{refund_id}{The id of the refund you want to update}

\item{args}{A list can contain:
\describe{
   \item{metadata}{\emph{optional} Any metadata in a list to change for refund.}
}}
}
\value{
A data frame with the updated refund info if successful.
}
\description{
Update a refund with any metadata you want to include about the refund.
}

